% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processes.R
\docType{data}
\name{CORES}
\alias{CORES}
\title{detect the number of CPU cores on the current host}
\format{
An object of class \code{integer} of length 1.
}
\usage{
CORES
}
\description{
detect the number of CPU cores on the current host
}
\keyword{datasets}
