% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processes.R
\docType{data}
\name{eo_ml_datacube}
\alias{eo_ml_datacube}
\title{return object from ml pedict process}
\format{
An object of class \code{list} of length 2.
}
\usage{
eo_ml_datacube
}
\description{
return object from ml pedict process
}
\keyword{datasets}
