% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processes.R
\docType{data}
\name{ml_fit_class_random_forest}
\alias{ml_fit_class_random_forest}
\title{fit random forest classification}
\format{
An object of class \code{Process} (inherits from \code{R6}) of length 12.
}
\usage{
ml_fit_class_random_forest
}
\description{
fit random forest classification
}
\keyword{datasets}
