% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Collection-class.R
\docType{data}
\name{SENTINEL_2_L2A}
\alias{SENTINEL_2_L2A}
\title{collections:
sentinel-s2-l2a}
\format{
An object of class \code{Collection} (inherits from \code{R6}) of length 13.
}
\usage{
SENTINEL_2_L2A
}
\description{
collections:
sentinel-s2-l2a
}
\keyword{datasets}
